+++ Demo/DemoTests/EndpointSpec.swift
@@ -45,14 +45,14 @@ class EndpointSpec: QuickSpec {
                 expect(newEndpoint.URL).to(equal(endpoint.URL))
                 expect(newEndpoint.method).to(equal(endpoint.method))
                 expect(newEndpoint.parameterEncoding).to(equal(endpoint.parameterEncoding))
-                expect(newEndpoint.httpHeaderFields.count).to(equal(endpoint.httpHeaderFields.count))
+                expect(newEndpoint.httpHeaderFields?.count).to(equal(endpoint.httpHeaderFields?.count))
             }
             
             it("returns a new endpoint for endpointByAddingHTTPHeaderFields") {
                 let agent = "Zalbinian"
                 let newEndpoint = endpoint.endpointByAddingHTTPHeaderFields(["User-Agent": agent])
                 
-                let newEndpointAgentObject: AnyObject? = newEndpoint.httpHeaderFields["User-Agent"]
+                let newEndpointAgentObject: AnyObject? = newEndpoint.httpHeaderFields?["User-Agent"]
                 let newEndpointAgent = newEndpointAgentObject as? String
                 // Make sure our closure updated the sample response, as proof that it can modify the Endpoint
                 expect(newEndpointAgent).to(equal(agent))
@@ -75,14 +75,14 @@ class EndpointSpec: QuickSpec {
                 expect(newEndpoint.URL).to(equal(endpoint.URL))
                 expect(newEndpoint.method).to(equal(endpoint.method))
                 expect(newEndpoint.parameters?.count).to(equal(endpoint.parameters?.count))
-                expect(newEndpoint.httpHeaderFields.count).to(equal(endpoint.httpHeaderFields.count))
+                expect(newEndpoint.httpHeaderFields?.count).to(equal(endpoint.httpHeaderFields?.count))
             }
             
             it("returns a correct URL request") {
                 let request = endpoint.urlRequest
                 expect(request.URL!.absoluteString).to(equal("https://api.github.com/zen"))
                 expect(NSString(data: request.HTTPBody!, encoding: 4)).to(equal("{\"Nemesis\":\"Harvey\"}"))
-                let titleObject: AnyObject? = endpoint.httpHeaderFields["Title"]
+                let titleObject: AnyObject? = endpoint.httpHeaderFields?["Title"]
                 let title = titleObject as? String
                 expect(title).to(equal("Dominar"))
             }